主键统一使用long类型

+++ zikucar-java/think-dao/src/main/java/com/think/common/BaseEntity.java
@@ -13,5 +13,5 @@ import java.io.Serializable;
 @Data
 public class BaseEntity implements Serializable {
     @TableId(type = IdType.AUTO)
-    private Integer pkId;
+    private Long pkId;
 }
